+++ src/Transformer/MultipleTransformer.php
@@ -35,6 +35,9 @@ final class MultipleTransformer implements TransformerInterface, DependentTransf
         Type::BUILTIN_TYPE_ITERABLE => 'is_iterable',
     ];
 
+    /**
+     * @param array<array{transformer: TransformerInterface, type: Type}> $transformers
+     */
     public function __construct(
         private readonly array $transformers,
     ) {
@@ -50,6 +53,18 @@ final class MultipleTransformer implements TransformerInterface, DependentTransf
             new Stmt\Expression(new Expr\Assign($output, $input)),
         ];
 
+        /*
+         * In case of the source type can be mixed we need to check the type before doing the transformation.
+         *
+         *  if (is_bool($input)) {
+         *     $output = $input;
+         *  }
+         *
+         *  if (is_int($input)) {
+         *     $output = (bool) $input;
+         *  }
+         *
+         */
         foreach ($this->transformers as $transformerData) {
             $transformer = $transformerData['transformer'];
             $type = $transformerData['type'];